Add experimental exit test support.
This PR adds a new kind of expectation called an "exit test", "death test",
"expected crasher", etc. It can be invoked by writing:

```swift
  fatalError("How dare you!")
}
```

It works by spinning up a second process that re-runs the same test function.
The parent process awaits the return of the child process, while the child
process executes the closure passed to `#expect()`.

This is similar to calling `fork()`, however that function is fundamentally
unusable on Darwin and unavailable on Windows, so I've pursued a different
implementation here that works across macOS, Linux, and Windows.

There are some constraints to using exit tests that are documented, but which
are not enforced at compile time. swift-syntax's `lexicalContext` should allow
us to do general checks for these constraints at compile time.

> [!WARNING]
> This is an experimental feature and is unlikely to ship in this form. It is
> not compatible with `XCTestScaffold` and can _only_ be used with the Swift 6
> toolchain/SwiftPM.

Resolves #157.
(Sort of.)