Only clear exit test ID env var upon successful lookup (#1226)
This changes the logic in `ExitTest.findInEnvironmentForEntryPoint()` to
only clear the environment variable containing the ID of the exit test
to run if an exit test is successfully located.

### Motivation:

This ensures that if a tool integrates with the testing library and has
both its own built-in copy of the library but also may call into the ABI
entry point, both usage scenarios can successfully look up the exit
test. Without this fix, if an earlier attempt to look up an exit test
fails, the environment variable would be cleared which prevents a
subsequent lookup attempt from succeeding.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.