Propagate `SkipInfo` to child tasks when a test is cancelled. (#1298)
This PR ensures that, when a test is cancelled, its test cases report
the same `SkipInfo` structure as they are recursively cancelled. For
instance:

```swift
@Test(arguments: [a ... z]) func foo(arg: T) throws {
  try Test.cancel("Message")
}
```

Prior to this change, the `.testCaseCancelled` events would be posted
without the user-supplied comment `"Message"` and each one would need to
construct a source context by capturing a backtrace from deep within
Swift Testing and the stdlib.

With this change in place, said events post with the same comment and
source context as is associated with the initial `.testCancelled` event.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.