Replace `wchar_t` with `CWideChar`. (#1412)
Previously, the definition of `CWideChar` was incorrect on Windows and
referred to a 32-bit type instead of a 16-bit type. This has been
corrected, so we can use Swift's canonical name for the type instead of
C's `wchar_t`.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.