Add qualification status for supported platforms in the README (#1424)
We've been able to build for more platforms in our CI, so update the
README to reflect this triumph. Explain how the different qualification
levels differ, and focus on build + test for all platforms in the
future.

### Modifications:

- Update platform support table
- Remove misc extra spaces in the README (I ran `prettier` over it)

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.